+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+-   [packages/core]: tests for scales and motion
 -   [packages/band]: components `Schedule` and `Schedules` for schedule / Gantt charts
 -   [examples]: examples of line charts with time scales
 -   [packages/core]: support for time scales